+++ inbm/Changelog.md
@@ -9,6 +9,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
  - RTC 509991 Updated docker stats dashboard to display units for readability.
  - HSD 15010649794 - INBC POTA/SOTA release date incorrect, SOTA schema change
  - HSD 15010868044 - Remove checking for isfile on destination before moving file
+ - HSD 15010868047 - "Unsupported OS" when running POTA via inbc in CentOS container
 
 ## 3.0.9 - 2022-03-08
 NOTE: This release has not yet been security tested.